chanegd the integer field in the model to big integer field to allow more digits

+++ front_end/models.py
@@ -83,7 +83,7 @@ class Area(models.Model):
     ship = models.ForeignKey(Ship, on_delete=models.CASCADE)
     area_name = models.CharField(max_length=200)
     scans = models.IntegerField()
-    point_cloud_size = models.IntegerField(default=0)
+    point_cloud_size = models.BigIntegerField(default=0)
     raw_size = models.DecimalField(max_digits=5, decimal_places=2, default=0)
     processed_size = models.DecimalField(max_digits=5, decimal_places=2, default=0)
     exported_size = models.DecimalField(max_digits=5, decimal_places=2,default=0)
